Fixed images more, fixed width/height of canvas to be a bit tighter. Fixed terrible movement bug!!! Used wrong time component and got worse as you play. Now much more precise, but doesn't feel exactly the way I want ....
touched up some graphics.

+++ Jigsaw/Jigsaw/Jigsaw/Canvas.cs
@@ -17,7 +17,7 @@ namespace Jigsaw
         //width of the surrounding border
         private const float _percentage = 9.0f / 64.0f;
 
-        internal void setSize(int width, int height)
+        internal void setSize(float width, float height)
         {
             _size.X = width * (1.0f + _percentage * 2);
             _size.Y = height * (1.0f + _percentage * 2);
